Fix in-place sed invocations
From memory, this is one of those discrepancies between BSD and GNU `sed`. Basically, BSD's `-i` option expects an extension with which `sed` will keep backups of the modified files. It needs to be explicitly passed a zero-length value if we want it to not produce backup files. What was happening before was that `sed` was consuming the next flag, `-e`, as the extension.